repackage sructure, add podcast list initial classes  + Test will be added
