--- app/src/main/java/com/radiofarda/istgah/istgaah/Tools/SoundCloudParser.java
@@ -1,14 +1,0 @@
-package com.radiofarda.istgah.istgaah.tools;
-
-import org.json.JSONObject;
-
-/**
- * Created by Farshad on 2/5/17.
- * Class to parse from sound cloud
- */
-
-public class SoundCloudParser {
-    public boolean parse(JSONObject jsonObject) {
-        return true;
-    }
-}
