+++ app/src/main/java/com/radiofarda/istgah/network/podcast/PodcastsFetcherCallback.java
@@ -1,4 +1,4 @@
-package com.radiofarda.istgah.istgaah.tools;
+package com.radiofarda.istgah.network.podcast;
 
 import java.util.List;
