+++ app/src/main/java/com/radiofarda/istgah/network/podcast/ProgramList.java
@@ -1,5 +1,5 @@
 
-package com.radiofarda.istgah.istgaah.tools;
+package com.radiofarda.istgah.network.podcast;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
