+++ app/src/main/java/com/radiofarda/istgah/utils/CommonUtils.java
@@ -1,0 +1,5 @@
+package com.radiofarda.istgah.utils;
+
+public class CommonUtils {
+
+}
