+++ app/src/main/res/layout/fragment_podcasts.xml
@@ -6,11 +6,11 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".fragments.PlaceholderFragment">
+    tools:context=".fragments.PodcastsFragment">
 
-    <TextView
-            android:id="@+id/section_label"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
+    <ListView xmlns:android="http://schemas.android.com/apk/res/android"
+              android:id="@+id/podcastList"
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content" />
 </RelativeLayout>
