--- app/src/test/java/com/radiofarda/istgah/istgaah/tools/PodcastsFetcherTest.java
@@ -1,19 +1,0 @@
-package com.radiofarda.istgah.istgaah.tools;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Created by Farshad Sheykhi on 2/5/17.
- * Test podcasts fetcher
- */
-
-public class PodcastsFetcherTest {
-    @Test
-    public void testParser() throws Exception {
-        PodcastsFetcherTest subject = new PodcastsFetcherTest();
-
-        assertEquals(true, true);
-    }
-}
